+++ NEWS.md
@@ -6,9 +6,11 @@ Find out the main highlights of each update.
 
 Fixes:
   * Fix crash if `process_mr()` is run without any SMR method requested.
+  * Export epoc functions.
 
 Enhancements:
   * Improve column names on output of `extract_epoc()`.
+  * New `calc_single_slope()` function.
 
 ## pyroresp 0.1.0